pyproject.toml
[project]
name = "htr-ocr"
version = "0.1.0"
description = "Diploma project: Transformers for Handwritten Text Recognition"
readme = "README.md"
requires-python = ">=3.10"
authors = [{ name = "student" }]

dependencies = [
  "hydra-core>=1.3.2",
  "omegaconf>=2.3.0",
  "fire>=0.6.0",
  "mlflow>=2.12.0",
  "pandas>=2.0.0",
  "pyarrow>=14.0.0",
  "pillow>=10.0.0",
  "tqdm>=4.66.0",
  "rich>=13.7.0",
]

[project.optional-dependencies]
train = [
  "torch==2.5.1+cu121",
  "torchvision==0.20.1+cu121",
  "torchaudio==2.5.1+cu121",
  "numpy>=1.26.0",
]

[project.scripts]
htr = "htr_ocr.cli:main"

[tool.uv]
index-strategy = "unsafe-best-match"

[[tool.uv.index]]
name = "pypi"
url = "https://pypi.org/simple"

[[tool.uv.index]]
name = "pytorch-cu121"
url = "https://download.pytorch.org/whl/cu121"

[build-system]
requires = ["setuptools>=68"]
build-backend = "setuptools.build_meta"

[tool.setuptools]
package-dir = {"" = "src"}

[tool.setuptools.packages.find]
where = ["src"]
